Removed hasVariants
Seemed just... out of place. For those interested, it basically checked
if the current block had other data values that actually did something.
This messed up mods (for example, if a mod had a special type of
bedrock, mcedit would replace it as general bedrock) and caused issues
in the blockpicker.