1.8 -> 1.9 Block name changes. Part 1: Blocks 1-98
This name update change is taking all the following into account:
1. Creative mode inventory names (in English).
2. Probable creative mode inventory names for blocks IF they existed in
the creative mode inventory. We have a bit more of an inventory in
mcedit. This usually just uses  "minecraft:block_name" in a "Block Name"
format to give the creative mode version.
3. All potential search terms (including blockstates, old names, and
other stuff)
4. Best display names in the Block Picker UI
5. Best display names elsewhere (like when alt-hovering your mouse over
a block)

4 & 5 will mostly have to do with when I used "aka". With "aka" it shows
up in the block picker pane, but doesn't show up on alt-hover block
info.

There are a few exceptions to these rules I don't feel like listing. But
if you see any weirdness to these guidelines, I probably meant to do it
on purpose and had a reasoning, but feel free to ask me why anyway and
we can debate it. My reasoning can be flawed.
Ahh who are we kidding my reasoning is never flawed, maybe just
incomplete!