Update de_DE.trn
Changed many translations and hopefully improved them. Half of them were spelling mistakes, the other ones were mostly grammar mistakes. I changed "Klon" to "Kopie", because it was named like this already at some translations, I hope I didn't messed it up. Changed "Bewohner" to "Dorfbewohner" because this is more correct (http://minecraft-de.gamepedia.com/Dorfbewohner). Changed "Kennzeichen" to "Tags". Changed to formal speech (were it wasn't) because the first half of the translation file was like this ("du" -> "Sie"). 
In line 1580 (t737) is a space "{0: n}", I am not quite sure if this is intended. And there are some translation missing, for example the update text and may I suggest using different translation files for the inbuild filters?