Update de_DE.trn
Change 696 because it had one space too much.
Added translation for 831 - 834 (note that these translations may partwise be not so good)
835 is still missing because I don't know the context.

At the moment there are the translations "Tool" and "Werkzeug" for the English "tool". Maybe we should use only one, rather "Werkzeug".
Changed 273, I am using "fragen" here for "prompt". I know that is not the correct translation, but I hope/think it fits pretty well.