Changed window handling system.
The new system (based on a class to communicate with the OS) is the default one for Linux platforms.
It handles the size, position and state of the window, exposing a unified and agnostic API.
It can, on Linux, re-maximize the window at startup if MCEdit was closed maximized.
When starting maximized, it restores the last used position and size.

A new command-line switch can be used to debug this: '--debug-wm'.

Four desktop environments are for now supported on Linux:

* cinnamon (X-Cinnamon as on Linux Mint)
* gnome (aka gnome-fallback on some Ubuntu flavors)
* kde (used by openSuSE)
* unity (mainly Ubuntu)

Tests are needed on the other environment. Some of them may be compatible with the actual code.

A Windows prototype has been implemented too, but not the the native OS.
To activate the new handling system on Windows, launch MCEdit from a console with the switch '--debug-wm'.

Badly, nothing has been done for OSX. Need someone with this OS...