Fixed #615
Efforts has been made to get the blocks and item translation directly from the game.

However it is not perfect and we lack some strings.
To translate them, separated files were created in the `Items` folder.

Two translations are for now available (`fr_FR` and `sv_SE`).
Feel free to create one using the `en_GB` file as a basis :smile: