Fix for 'KeyError(XX,)' when loading/saving
Relative to issue #664, and some other ones.

The entities present in worlds and not yet referenced in MCEdit will no more display an error.
They will be created 'on the fly' with a 'default' name ('Entity XX').
A warning massage saying this unknown entity was found is displayed in the console and written to the log file.