Reworked ALBOW layout widget
The Row and Column ALBOW widgets now accept new values as 'expand' keyword argument.

Setting this keword argument to 'h' will expand all the sub-widgets horizontaly.
Setting it to 'v' will expand the sub-widgets verticaly.

The value can also be a combination of both 'h' and 'v'.

The expansion is made according to the greatest sub-widget width and/or height.