Fixed an issue with waypoints and dimensions
When the option `Save Camera position on world close` was enabled, changing world dimension was not possible: the view appeared to be sticked to the overworld.
By disabling this waypoints option and moving (or deleting) the related file, it was possible to switch between overworld, Neteher and The End again.

Now, the dimension can be changed, but when reloading a world, the view will be in the overworld.

Need more work :wink: