Implemented new feature activable with CLI and changed OpenGL debug mode CLI option
The OpenGL debug mode command line option has been changed from '-debug' to '--debug-ogl'.

New features activable whit CLI

The idea is to be able to implement and test/use new code without stripping off the current one.
These features/new code will be in the released stuff, but unavailable until explicitly requested.

The new features which are under development can be enabled using the 'new_features.def' file.
This file is a plain text file with one feature to enable a line.
The file is parsed and each feature is added to the builtins using the pattern 'mcenf_<feature>'.
The value for these builtins is 'True'.
Then, in the code, just check if the builtins has the key 'mcenf_<feature>' to use the new version of the code:

```
def foo_old():
    # Was 'foo', code here is the one used unless the new version is wanted.
    [...]

def foo_new():
    # This is the new version of the former 'foo' (current 'foo_old').
    [...]

if __builtins__.get('mcenf_foo', False):
    foo = foo_new
else:
    foo = foo_old

```