Implemented debug mode for PE support.
Use the `--debug-pe` CLI option to enable ti mode.

It uses the full Python NBT support because we need to get some information about the errors.
It also writes a file called `dump_pe.txt`.
Be careful with this debug mode, a lot of data is written in the file!