Fix for player creation.
The player creation (using the Player panel 'Add' button) was broken because it was recieving a thread object instead of usable data.
The solution may not be perfect, but no warning popup concerning bad data type is displayed, and the player is added to the list.
Need more tests with 'real' player.

Alos commented some debug strings displayed in the console. Some may have be left...