Made leveldb.py behave like current leveldb_mcpe
Tweaking 'compression' in leveldb.py and adding some ugly bypass make the pure python module work like the C/PythonBoost library.
Pocket words before 1.0 can be read, modified and saved (changes are visible when reloading world in MCEdit).
Pocket worlds 1.0+ can be loaded and changed, but the changes are not visible when reloading in MCEdit.