Reworked mob spawners for PE
The entity IDs lookup has been changed and specific spawner data updating code has been implemented.

Some weird stuff happens with displayed entity names, but it does affect the full numerical ID to be used when saving the world.
I gues it's something to do with the hardcoded entities names and IDs which are used in the 'Entity' classes.