Basic .jar mod file finding
The .jar is retrieved using the mcmod.info content.
No version check for the mod implemented for now. (Some mod do not report the correct version in the level.dat).
Searched directories ar the 'mods' directories in MCEdit one and in the world parent directory.
ModLoader called with the .jar name(s) to initialize the data we need.