Partially fixed pistons
Pistons facing horisontally needed their direction changing.
Removed higher data values that seem to have been coppied from Java but are not used in Bedrock.
Added a 50:50 normal:sticky texture for the piston extension since the same block is shared for both. Ideally a custom renderer would access the connected block to get the correct texture but this is beyond my knowledge