fixed sub-chunks not saving if they didn't already exist
Previously for data to be written to disk the chunk and sub-chunk had to exist when MCedit was loaded. I wrote an MCedit filter to create them but thought I would look into fixing this internally so the filter was not needed.
The sub-chunk key only needs to exist in the database if it contains blocks that are not just air so the key is deleted if there is only air.