Fixed bug saving blocks not in the JSON files
The old code was looking in the blocksByID dictionary. If the id:data pair was not defined in the JSON files but another data value was defined it would not exist so would be replaced with air. This allows any data value to be saved regardless of its presence in the JSON files. (Note: This allows you to save invalid states which at best the game will replace with air or at worst cause the game to crash.  This is however the cost to allow valid states that are not in the JSON files)