Numerious performance improvements (#935)
* masively simplify lighting calculations

* significantly faster chunk saving for Bedrock V8 subchunks

this function previously took about 0.5 seconds to run for all the subchunks in a chunk. It now takes about 0.005 seconds (100 times faster)

* slightly improved getAllPlayerData speed

for a large world reduced the runtime of getAllPlayerData from 10.6 seconds to 6.7 seconds

* Revert "slightly improved getAllPlayerData speed"

This reverts commit 247d269bdbde569f0bb5a1eb091a45325b419262.

* removed unnecessary while loop and changed if requirement

the length of the terrain should be based on num_of_storages not the length of the actual data. If further storages are made accessible a for loop can be added

* Improved the efficiency of getAllChunks

Previously this function iterated through literally every key in the database and tested each one if it was the base chunk key. Now it skips over all the keys with the same first 8 characters (the location marker). This seems to about half the runtime of the function

* Converted allChunks to use a set rather than a list

In large worlds the list returned by allChunks is very long and takes a little while to test if a value is contained within. Sets take fractionally longer to build but are far quicker at testing if a value is contained within them

* get => idStr and merged into one for loop

Performance optimisation. The get method was quite slow so has been replaced with just looking for the id in the idStr list. 4.47ms per call down to 0.72ms per call

* significantly sped up chunk creation (__init__)

Changed to use other existing data sets that can be iterated over quicker
Old code : 15.6ms per call
New code : 0.76ms per call

* Removed line intended to be removed in the last commit and old code

The top line was supposed to be part of the previous commit
The lower block is old redundant code

* Fixed bug with getAllChunks

getAllChunks was catching keys not indended to be caught due to db.Get working in a different way to as thought

* Removed line indented to be part of the last commit

* Removed .lower() as some string names contain capitals

* Fixed bug saving blocks not in the JSON files

The old code was looking in the blocksByID dictionary. If the id:data pair was not defined in the JSON files but another data value was defined it would not exist so would be replaced with air. This allows any data value to be saved regardless of its presence in the JSON files. (Note: This allows you to save invalid states which at best the game will replace with air or at worst cause the game to crash.  This is however the cost to allow valid states that are not in the JSON files)

* Moved light calculation to chunk setup