Fixed the negative time in saving and accidentally also the chunk grid
This function was not defined that gets the number of chunks in the world to give an approximate time until saving is complete. This has been improved to return the number of loaded chunks to give a better time estimate.

This accidentally also caused the chunk grid to work so I guess there is that as well