added the posibility for dimension support in the code for Bedrock edition
To change dimension you will need to be running from source so that you can edit the source code. First fully close MCedit and then open up the pymclevel/leveldbpocket.py file in a text editor and uncomment one of the first three lines related to the dimension you want to edit and comment the others. Then open up the world that you want to and that dimension will be loaded. Edit away.

Note: some filters will edit the overworld no matter which dimension you have open so be careful and backup worlds before starting